FRW-8639 updated readme and script

+++ otel_integrator.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Default values
-project_namespace="Pyz"
+PROJECT_NAMESPACE="Pyz"
 IMAGE_FILE=""
 INSTALL_FILE="config/install/docker.yml"
 DEPLOY_FILE="deploy.yml"
@@ -12,11 +12,11 @@ display_help() {
   echo "Usage: $0 [options]"
   echo ""
   echo "Options:"
-  echo "  --project-namespace <namespace>  Set the project namespace (default: Pyz)"
-  echo "  --image-file <file>              Specify the image file (default: same as --deploy-file if not provided)"
-  echo "  --install-file <file>            Path to the install configuration file (default: config/install/docker.yml)"
-  echo "  --deploy-file <file>             Path to the deploy configuration file (default: deploy.yml)"
-  echo "  --base-image <image>             Base Docker image (default: volhovm/spryker-8.3-alpine-3.20)"
+  echo "  --project-namespace <namespace>  Set the project namespace. Will be used to wire monitoring plugin and console command. (default: Pyz)"
+  echo "  --image-file <file>              Specify the image file that includes your PHP image information. You need to use it if your main deploy file doesn't have it (default: same as --deploy-file if not provided)"
+  echo "  --install-file <file>            Path to the install configuration file. Will be updated with hook generator command. (default: config/install/docker.yml)"
+  echo "  --deploy-file <file>             Path to the deploy configuration file. If --image-file is not provided this file will be used both for PHP image update and for booting application. If --image-file is provided, only to boot application. (default: deploy.yml)"
+  echo "  --base-image <image>             Base PHP Docker image. Image MUST be built with all required extensions. (default: volhovm/spryker-8.3-alpine-3.20)"
   echo "  --help                           Display this help message"
   echo ""
   echo "Example:"
@@ -36,7 +36,7 @@ check_status() {
 while [[ $# -gt 0 ]]; do
   case $1 in
     --project-namespace)
-      project_namespace="$2"
+      PROJECT_NAMESPACE="$2"
       shift 2
       ;;
     --image-file)
@@ -65,7 +65,7 @@ while [[ $# -gt 0 ]]; do
   esac
 done
 
-# Set IMAGE_FILE to DEPLOY_FILE if IMAGE_FILE is not explicitly provided
+# Set value from DEPLOY_FILE to IMAGE_FILE if IMAGE_FILE is not explicitly provided. This means that your deploy file also includes image information.
 if [[ -z $IMAGE_FILE ]]; then
   IMAGE_FILE="$DEPLOY_FILE"
 fi
@@ -73,11 +73,11 @@ fi
 # Display configuration
 echo "Configuration summary:"
 echo "======================"
-echo "Project Namespace: $project_namespace"
+echo "Project Namespace: $PROJECT_NAMESPACE"
 echo "Image File: $IMAGE_FILE"
 echo "Install File: $INSTALL_FILE"
 echo "Deploy File: $DEPLOY_FILE"
-echo "Base Image: $BASE_IMAGE"
+echo "PHP Image: $BASE_IMAGE"
 echo "======================"
 
 # Ask for confirmation
@@ -127,6 +127,12 @@ adjustDeployFile() {
     local currentDir
     currentDir=$(pwd)
     cmd=( -i -I4 '.image.tag = "'$BASE_IMAGE'" | .image.php.enabled-extensions |= (.|select(. != null) + ["opentelemetry", "grpc", "protobuf"] | unique) // ["opentelemetry", "grpc", "protobuf"]' "$deployFile" )
+
+    if [[ ! -f "$deployFile" ]]; then
+      echo "Error: File $deployFile does not exist."
+      exit 1
+    fi
+
     echo "Execution command: docker run --rm -v $currentDir:/workdir mikefarah/yq --no-doc --indent 2 ${cmd[@]}"
     docker run --rm -v "$currentDir:/workdir" mikefarah/yq "${cmd[@]}"
     check_status $? "Failed to adjust deploy file: $deployFile"
@@ -151,9 +157,9 @@ installDependencies() {
 }
 
 registerPlugins() {
-    if [ -f "src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php" ]; then
+    if [ -f "src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php" ]; then
         # Check if the getMonitoringExtensions method exists
-        if ! grep -q "function getMonitoringExtensions" "src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php"; then
+        if ! grep -q "function getMonitoringExtensions" "src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php"; then
                 # If not present, add the method and the necessary use statements
                 runSedCommand '/^}$/i\
         /**\
@@ -165,28 +171,28 @@ registerPlugins() {
                 new OpentelemetryMonitoringExtensionPlugin(),\
             ];\
         }
-        ' src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php
+        ' src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php
         fi
 
         # Check if the plugins are already present
-        if ! grep -q "OpentelemetryMonitoringExtensionPlugin" "src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php"; then
+        if ! grep -q "OpentelemetryMonitoringExtensionPlugin" "src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php"; then
 
             # If not present, add them to the return array
             runSedCommand '/        return \[/a\
             \ \ \ \ new OpentelemetryMonitoringExtensionPlugin(),
-            ' src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php
+            ' src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php
         fi
 
-        if ! grep -Fq "use Spryker\\Service\\Opentelemetry\\Plugin\\OpentelemetryMonitoringExtensionPlugin;" "src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php"; then
-            runSedCommand "/^namespace ${project_namespace//\\/\\\\}\\\\Service\\\\Monitoring;/a\\
+        if ! grep -Fq "use Spryker\\Service\\Opentelemetry\\Plugin\\OpentelemetryMonitoringExtensionPlugin;" "src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php"; then
+            runSedCommand "/^namespace ${PROJECT_NAMESPACE//\\/\\\\}\\\\Service\\\\Monitoring;/a\\
 use Spryker\\\\Service\\\\Opentelemetry\\\\Plugin\\\\OpentelemetryMonitoringExtensionPlugin;
-            " "src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php"
+            " "src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php"
         fi
     else
         # If the file doesn't exist, create it
         # Create the directory if it doesn't exist
-        mkdir -p "src/$project_namespace/Service/Monitoring"
-        cat <<EOF > "src/$project_namespace/Service/Monitoring/MonitoringDependencyProvider.php"
+        mkdir -p "src/$PROJECT_NAMESPACE/Service/Monitoring"
+        cat <<EOF > "src/$PROJECT_NAMESPACE/Service/Monitoring/MonitoringDependencyProvider.php"
 <?php
 
 /**
@@ -194,7 +200,7 @@ use Spryker\\\\Service\\\\Opentelemetry\\\\Plugin\\\\OpentelemetryMonitoringExte
  * For full license information, please view the LICENSE file that was distributed with this source code.
  */
 
-namespace $project_namespace\Service\Monitoring;
+namespace $PROJECT_NAMESPACE\Service\Monitoring;
 
 use Spryker\Service\Monitoring\MonitoringDependencyProvider as SprykerMonitoringDependencyProvider;
 use Spryker\Service\Opentelemetry\Plugin\OpentelemetryMonitoringExtensionPlugin;
@@ -215,18 +221,18 @@ EOF
     fi
 
     # Check if the file exists
-    if [ -f "src/$project_namespace/Zed/Console/ConsoleDependencyProvider.php" ]; then
+    if [ -f "src/$PROJECT_NAMESPACE/Zed/Console/ConsoleDependencyProvider.php" ]; then
         # If the file exists, attempt to append the code to the $commands array
-        if ! grep -q "OpentelemetryGeneratorConsole" "src/$project_namespace/Zed/Console/ConsoleDependencyProvider.php"; then
+        if ! grep -q "OpentelemetryGeneratorConsole" "src/$PROJECT_NAMESPACE/Zed/Console/ConsoleDependencyProvider.php"; then
             runSedCommand '/$commands = \[/a\
             new OpentelemetryGeneratorConsole(),
-                ' src/$project_namespace/Zed/Console/ConsoleDependencyProvider.php
+                ' src/$PROJECT_NAMESPACE/Zed/Console/ConsoleDependencyProvider.php
 
                 # Add the use statement if it doesn't already exist
-                if ! grep -q "use Spryker\\Zed\\Opentelemetry\\Communication\\Plugin\\Console\\OpentelemetryGeneratorConsole;" "src/$project_namespace/Zed/Console/ConsoleDependencyProvider.php"; then
+                if ! grep -q "use Spryker\\Zed\\Opentelemetry\\Communication\\Plugin\\Console\\OpentelemetryGeneratorConsole;" "src/$PROJECT_NAMESPACE/Zed/Console/ConsoleDependencyProvider.php"; then
                     runSedCommand "
-                    /^namespace ${project_namespace//\\/\\\\}\\\\Zed\\\\Console;/a\\
-use Spryker\\\\Zed\\\\Opentelemetry\\\\Communication\\\\Plugin\\\\Console\\\\OpentelemetryGeneratorConsole;" src/$project_namespace/Zed/Console/ConsoleDependencyProvider.php
+                    /^namespace ${PROJECT_NAMESPACE//\\/\\\\}\\\\Zed\\\\Console;/a\\
+use Spryker\\\\Zed\\\\Opentelemetry\\\\Communication\\\\Plugin\\\\Console\\\\OpentelemetryGeneratorConsole;" src/$PROJECT_NAMESPACE/Zed/Console/ConsoleDependencyProvider.php
                 fi
         fi
     fi
